// src/lib/types.ts
export type AwesomeAPIData = {
  code: string;
  codein: string;
  name: string;
  high: string;
  low: string;
  varBid: string;
  pctChange: string;
  bid: string;
  ask: string;
  timestamp: string;
  create_date: string;
};

export type AwesomeAPIResponse = {
  [key: string]: AwesomeAPIData;
};